update fab for new server

+++ fabfile.py
@@ -1,7 +1,16 @@
 from fabric.api import *
-
 env.path = '/var/sites/newsconnector'
-env.hosts = ['ubuntu@newsworld.co.za']
+is_web = False
+
+
+def cloud():
+    env.hosts = ['ubuntu@162.243.76.49']
+
+
+def web():
+    global is_web
+    is_web = True
+    env.hosts = ['ubuntu@192.241.255.58']
 
 
 def push():
@@ -10,13 +19,16 @@ def push():
 
 
 def static():
-    with cd(env.path):
-        run('ve/bin/python %(path)s/newsconnector/manage.py collectstatic --noinput' % env)
+    if is_web:
+        with cd(env.path):
+            run('ve/bin/python %(path)s/newsconnector/manage.py '
+                'collectstatic --noinput' % env)
 
 
 def reload_g():
-    with cd(env.path):
-        run('kill -HUP `cat tmp/pids/newsconnector*.pid`')
+    if is_web:
+        with cd(env.path):
+            run('kill -HUP `cat tmp/pids/newsconnector*.pid`')
 
 
 def deploy():
@@ -26,13 +38,15 @@ def deploy():
 
 
 def restart():
-    sudo('supervisorctl restart newsconnector:')
+    if is_web:
+        sudo('supervisorctl restart newsconnector:')
 
 
 def restart_celery():
     with cd(env.path):
         sudo('supervisorctl stop celery')
-        run('ve/bin/python %(path)s/newsconnector/manage.py celery purge -f' % env)
+        run('ve/bin/python %(path)s/newsconnector/manage.py '
+            'celery purge -f' % env)
 
     with settings(warn_only=True):
         run('%(path)s/kill_workers.sh' % env)
